Move Miband3 component and names refactor
